requirements.txt
# Place your requirements here
python-telegram-bot[all]
